revert: add-habit-data PR #46
