--- src/api/habits/adapters/index.ts
@@ -1,3 +1,0 @@
-import jsonHabitServiceFactory from './json';
-
-export default jsonHabitServiceFactory();
