--- src/api/habits/index.ts
@@ -1,3 +1,0 @@
-import habitFactory from './adapters';
-
-export default habitFactory;
