--- src/api/habits/types.ts
@@ -1,10 +1,0 @@
-import { GoalWithHabitHistory } from '../../types/habits';
-
-export type HabitService = {
-  save?: () => {};
-  retrieve?: () => {};
-  get?: () => {};
-  set?: () => {};
-  addHabit: () => Promise<string>;
-  getHabits: () => Promise<GoalWithHabitHistory[]>;
-};
